Moved files out of algorithms folder
