--- openpharmacophore/tests/utils/test_align_ligands.py
@@ -1,3 +1,0 @@
-
-def test_align_set_of_ligands():
-    assert False, "Complete me!"
