+++ openpharmacophore/utils/bisection.py
